+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/meal/MealModule.kt
@@ -1,8 +1,8 @@
 package com.maksimowiczm.foodyou.feature.meal
 
-import com.maksimowiczm.foodyou.feature.meal.domain.ObserveMealsWithSummaryUseCase
-import com.maksimowiczm.foodyou.feature.meal.domain.ObserveMealsWithSummaryUseCaseImpl
-import com.maksimowiczm.foodyou.feature.meal.ui.card.MealsCardViewModel
+import com.maksimowiczm.foodyou.feature.meal.domain.ObserveMealsUseCase
+import com.maksimowiczm.foodyou.feature.meal.domain.ObserveMealsUseCaseImpl
+import com.maksimowiczm.foodyou.feature.meal.ui.card.MealsCardsViewModel
 import com.maksimowiczm.foodyou.feature.meal.ui.settings.MealsSettingsScreenViewModel
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.viewModelOf
@@ -10,8 +10,8 @@ import org.koin.dsl.bind
 import org.koin.dsl.module
 
 val mealModule = module {
-    factoryOf(::ObserveMealsWithSummaryUseCaseImpl).bind<ObserveMealsWithSummaryUseCase>()
+    factoryOf(::ObserveMealsUseCaseImpl).bind<ObserveMealsUseCase>()
 
+    viewModelOf(::MealsCardsViewModel)
     viewModelOf(::MealsSettingsScreenViewModel)
-    viewModelOf(::MealsCardViewModel)
 }
