+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/meal/ui/card/FoodListItem.kt
@@ -1,4 +1,4 @@
-package com.maksimowiczm.foodyou.feature.mealredesign.ui
+package com.maksimowiczm.foodyou.feature.meal.ui.card
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
