+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/meal/ui/card/MealCard.kt
@@ -1,176 +1,172 @@
 package com.maksimowiczm.foodyou.feature.meal.ui.card
 
-import androidx.compose.animation.AnimatedVisibilityScope
-import androidx.compose.animation.ExperimentalSharedTransitionApi
-import androidx.compose.animation.SharedTransitionScope
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeOut
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.animateContentSize
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.shrinkVertically
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.FilledIconButton
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
+import androidx.compose.material3.ListItem
+import androidx.compose.material3.ListItemDefaults
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.key
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import com.maksimowiczm.foodyou.core.domain.model.FoodWithMeasurement
+import com.maksimowiczm.foodyou.core.domain.model.MeasurementId
+import com.maksimowiczm.foodyou.core.ext.lambda
 import com.maksimowiczm.foodyou.core.ui.home.FoodYouHomeCard
-import com.maksimowiczm.foodyou.core.ui.motion.crossfadeIn
+import com.maksimowiczm.foodyou.core.ui.modifier.animatePlacement
+import com.maksimowiczm.foodyou.core.ui.theme.LocalNutrientsPalette
 import com.maksimowiczm.foodyou.core.ui.utils.LocalDateFormatter
-import com.maksimowiczm.foodyou.feature.meal.domain.MealWithSummary
-import com.maksimowiczm.foodyou.feature.meal.ui.card.MealCardTransitionSpecs.overlayClipFromCardToScreen
-import com.maksimowiczm.foodyou.feature.meal.ui.component.MealHeader
-import com.maksimowiczm.foodyou.feature.meal.ui.component.NutrientsLayout
-import foodyou.app.generated.resources.Res
-import foodyou.app.generated.resources.action_add
-import foodyou.app.generated.resources.em_dash
-import foodyou.app.generated.resources.en_dash
-import foodyou.app.generated.resources.headline_all_day
-import foodyou.app.generated.resources.unit_gram_short
+import com.maksimowiczm.foodyou.feature.meal.domain.Meal
+import foodyou.app.generated.resources.*
 import org.jetbrains.compose.resources.stringResource
 
-@OptIn(ExperimentalSharedTransitionApi::class)
+@OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
-internal fun SharedTransitionScope.MealCard(
-    animatedVisibilityScope: AnimatedVisibilityScope,
-    epochDay: Int,
-    meal: MealWithSummary,
-    isEmpty: Boolean,
-    totalCalories: Int,
-    totalProteins: Int,
-    totalCarbohydrates: Int,
-    totalFats: Int,
-    onMealClick: () -> Unit,
-    onAddClick: () -> Unit,
+internal fun MealCard(
+    meal: Meal,
+    onAddFood: () -> Unit,
     onLongClick: () -> Unit,
+    onEditMeasurement: (MeasurementId) -> Unit,
+    onDeleteEntry: (MeasurementId) -> Unit,
     modifier: Modifier = Modifier
 ) {
+    val nutrientsPalette = LocalNutrientsPalette.current
     val dateFormatter = LocalDateFormatter.current
+    val enDash = stringResource(Res.string.en_dash)
+    val allDayString = stringResource(Res.string.headline_all_day)
 
-    FoodYouHomeCard(
-        onClick = onMealClick,
-        modifier = modifier.sharedBounds(
-            sharedContentState = rememberSharedContentState(
-                key = MealCardTransitionKeys.MealContainer(
-                    mealId = meal.id,
-                    epochDay = epochDay
-                )
-            ),
-            animatedVisibilityScope = animatedVisibilityScope,
-            enter = MealCardTransitionSpecs.containerEnterTransition,
-            exit = MealCardTransitionSpecs.containerExitTransition,
-            resizeMode = SharedTransitionScope.ResizeMode.RemeasureToBounds,
-            clipInOverlayDuringTransition = OverlayClip(
-                animatedVisibilityScope.overlayClipFromCardToScreen()
-            )
-        ),
-        onLongClick = onLongClick
-    ) {
-        val headline = @Composable {
-            Text(
-                text = meal.name,
-                modifier = Modifier.sharedElement(
-                    sharedContentState = rememberSharedContentState(
-                        key = MealCardTransitionKeys.MealTitle(
-                            mealId = meal.id,
-                            epochDay = epochDay
-                        )
-                    ),
-                    animatedVisibilityScope = animatedVisibilityScope
-                )
-            )
-        }
-        val time = @Composable {
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(4.dp),
-                modifier = Modifier.sharedElement(
-                    sharedContentState = rememberSharedContentState(
-                        key = MealCardTransitionKeys.MealTime(
-                            mealId = meal.id,
-                            epochDay = epochDay
-                        )
-                    ),
-                    animatedVisibilityScope = animatedVisibilityScope
-                )
-            ) {
-                if (meal.isAllDay) {
-                    Text(
-                        text = stringResource(Res.string.headline_all_day),
-                        color = MaterialTheme.colorScheme.outline
-                    )
-                } else {
-                    val enDash = stringResource(Res.string.en_dash)
-
-                    Text(
-                        text = remember(enDash, meal, dateFormatter) {
-                            buildString {
-                                append(dateFormatter.formatTime(meal.from))
-                                append(" ")
-                                append(enDash)
-                                append(" ")
-                                append(dateFormatter.formatTime(meal.to))
-                            }
-                        },
-                        color = MaterialTheme.colorScheme.outline
-                    )
-                }
+    val timeString = remember(dateFormatter, meal, enDash, allDayString) {
+        if (meal.isAllDay) {
+            allDayString
+        } else {
+            buildString {
+                append(dateFormatter.formatTime(meal.from))
+                append(" $enDash ")
+                append(dateFormatter.formatTime(meal.to))
             }
         }
+    }
 
-        val caloriesLabel = @Composable {
+    FoodYouHomeCard(
+        modifier = modifier,
+        onClick = {},
+        onLongClick = onLongClick
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(16.dp)
+        ) {
             Text(
-                text = if (isEmpty) {
-                    stringResource(Res.string.em_dash)
-                } else {
-                    totalCalories.toString()
-                }
+                text = meal.name,
+                style = MaterialTheme.typography.headlineMediumEmphasized,
+                color = MaterialTheme.colorScheme.onSurface
             )
-        }
-        val proteinsLabel = @Composable {
             Text(
-                text = if (isEmpty) {
-                    stringResource(Res.string.em_dash)
-                } else {
-                    "$totalProteins " + stringResource(Res.string.unit_gram_short)
-                }
+                text = timeString,
+                style = MaterialTheme.typography.labelLarge,
+                color = MaterialTheme.colorScheme.onSurfaceVariant
             )
-        }
-        val carbohydratesLabel = @Composable {
-            Text(
-                text = if (isEmpty) {
-                    stringResource(Res.string.em_dash)
-                } else {
-                    "$totalCarbohydrates " + stringResource(Res.string.unit_gram_short)
-                }
-            )
-        }
-        val fatsLabel = @Composable {
-            Text(
-                text = if (isEmpty) {
-                    stringResource(Res.string.em_dash)
-                } else {
-                    "$totalFats " + stringResource(Res.string.unit_gram_short)
-                }
-            )
-        }
 
-        val actionButton = @Composable {
-            with(animatedVisibilityScope) {
+            Spacer(Modifier.height(16.dp))
+
+            FoodContainer(
+                food = meal.food,
+                onEditMeasurement = onEditMeasurement,
+                onDeleteEntry = onDeleteEntry,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clip(MaterialTheme.shapes.medium)
+                    .animateContentSize(MaterialTheme.motionScheme.defaultSpatialSpec())
+            )
+
+            AnimatedVisibility(
+                visible = meal.food.isNotEmpty(),
+                enter = expandVertically(
+                    animationSpec = MaterialTheme.motionScheme.defaultSpatialSpec()
+                ),
+                exit = shrinkVertically(
+                    animationSpec = MaterialTheme.motionScheme.defaultSpatialSpec()
+                )
+            ) {
+                Spacer(Modifier.height(16.dp))
+            }
+
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.spacedBy(16.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                ValueColumn(
+                    label = stringResource(Res.string.unit_kcal),
+                    value = meal.calories,
+                    suffix = null,
+                    color = MaterialTheme.colorScheme.onSurface
+                )
+
+                ValueColumn(
+                    label = stringResource(Res.string.nutriment_proteins_short),
+                    value = meal.proteins,
+                    suffix = stringResource(Res.string.unit_gram_short),
+                    color = nutrientsPalette.proteinsOnSurfaceContainer
+                )
+
+                ValueColumn(
+                    label = stringResource(Res.string.nutriment_carbohydrates_short),
+                    value = meal.carbohydrates,
+                    suffix = stringResource(Res.string.unit_gram_short),
+                    color = nutrientsPalette.carbohydratesOnSurfaceContainer
+                )
+
+                ValueColumn(
+                    label = stringResource(Res.string.nutriment_fats_short),
+                    value = meal.fats,
+                    suffix = stringResource(Res.string.unit_gram_short),
+                    color = nutrientsPalette.fatsOnSurfaceContainer
+                )
+
+                Spacer(Modifier.weight(1f))
+
                 FilledIconButton(
-                    onClick = onAddClick,
-                    modifier = Modifier
-                        .renderInSharedTransitionScopeOverlay(zIndexInOverlay = 1f)
-                        .animateEnterExit(
-                            enter = crossfadeIn(),
-                            exit = fadeOut(tween(50))
-                        )
+                    onClick = onAddFood,
+                    shape = MaterialTheme.shapes.medium
                 ) {
                     Icon(
                         imageVector = Icons.Default.Add,
@@ -179,37 +175,199 @@ internal fun SharedTransitionScope.MealCard(
                 }
             }
         }
+    }
+}
 
-        MealHeader(
-            headline = headline,
-            time = time,
-            modifier = Modifier.padding(16.dp),
-            nutrientsLayout = {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    NutrientsLayout(
-                        caloriesLabel = caloriesLabel,
-                        proteinsLabel = proteinsLabel,
-                        carbohydratesLabel = carbohydratesLabel,
-                        fatsLabel = fatsLabel,
-                        modifier = Modifier.sharedElement(
-                            sharedContentState = rememberSharedContentState(
-                                key = MealCardTransitionKeys.MealNutrients(
-                                    mealId = meal.id,
-                                    epochDay = epochDay
-                                )
-                            ),
-                            animatedVisibilityScope = animatedVisibilityScope
-                        )
-                    )
+@OptIn(ExperimentalMaterial3ExpressiveApi::class)
+@Composable
+private fun FoodContainer(
+    food: List<FoodWithMeasurement>,
+    onEditMeasurement: (MeasurementId) -> Unit,
+    onDeleteEntry: (MeasurementId) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier,
+        verticalArrangement = Arrangement.spacedBy(2.dp)
+    ) {
+        food.forEach { foodWithMeasurement ->
+            key(foodWithMeasurement.measurementId) {
+                FoodContainerItem(
+                    foodWithMeasurement = foodWithMeasurement,
+                    onEditMeasurement = onEditMeasurement,
+                    onDeleteEntry = onDeleteEntry,
+                    modifier = Modifier.animatePlacement()
+                )
+            }
+        }
+    }
+}
 
-                    Spacer(Modifier.weight(1f))
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun FoodContainerItem(
+    foodWithMeasurement: FoodWithMeasurement,
+    onEditMeasurement: (MeasurementId) -> Unit,
+    onDeleteEntry: (MeasurementId) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    var showBottomSheet by rememberSaveable { mutableStateOf(false) }
+    val coroutineScope = rememberCoroutineScope()
 
-                    actionButton()
+    if (showBottomSheet) {
+        val sheetState = rememberModalBottomSheetState()
+
+        ModalBottomSheet(
+            onDismissRequest = { showBottomSheet = false },
+            sheetState = sheetState
+        ) {
+            BottomSheetContent(
+                food = foodWithMeasurement,
+                onEdit = coroutineScope.lambda {
+                    onEditMeasurement(foodWithMeasurement.measurementId)
+                    sheetState.hide()
+                    showBottomSheet = false
+                },
+                onDelete = coroutineScope.lambda {
+                    sheetState.hide()
+                    onDeleteEntry(foodWithMeasurement.measurementId)
+                    showBottomSheet = false
                 }
+            )
+        }
+    }
+
+    FoodListItem(
+        foodWithMeasurement = foodWithMeasurement,
+        modifier = modifier.clickable { showBottomSheet = true },
+        color = MaterialTheme.colorScheme.surfaceVariant,
+        contentColor = MaterialTheme.colorScheme.onSurface
+    )
+}
+
+@Composable
+private fun ValueColumn(
+    label: String,
+    value: Int,
+    suffix: String?,
+    color: Color,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(4.dp)
+    ) {
+        CompositionLocalProvider(
+            LocalContentColor provides color,
+            LocalTextStyle provides MaterialTheme.typography.labelMedium
+        ) {
+            Text(
+                text = label,
+                style = MaterialTheme.typography.labelMedium
+            )
+
+            Text(
+                text = if (value == 0) {
+                    stringResource(Res.string.em_dash)
+                } else {
+                    value.toString() + (suffix?.let { " $suffix" } ?: "")
+                }
+            )
+        }
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun BottomSheetContent(
+    food: FoodWithMeasurement,
+    onEdit: () -> Unit,
+    onDelete: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    var showDeleteDialog by rememberSaveable { mutableStateOf(false) }
+
+    if (showDeleteDialog) {
+        DeleteDialog(
+            onDismissRequest = { showDeleteDialog = false },
+            onDeleteEntry = {
+                onDelete()
+                showDeleteDialog = false
             }
         )
     }
+
+    Column(
+        modifier = modifier
+    ) {
+        FoodListItem(
+            foodWithMeasurement = food,
+            color = Color.Transparent,
+            contentColor = MaterialTheme.colorScheme.onSurface
+        )
+        HorizontalDivider(Modifier.padding(horizontal = 16.dp))
+        ListItem(
+            headlineContent = {
+                Text(stringResource(Res.string.action_edit_entry))
+            },
+            modifier = Modifier.clickable { onEdit() },
+            leadingContent = {
+                Icon(
+                    imageVector = Icons.Default.Edit,
+                    contentDescription = null
+                )
+            },
+            colors = ListItemDefaults.colors(
+                containerColor = Color.Transparent
+            )
+        )
+        ListItem(
+            headlineContent = {
+                Text(stringResource(Res.string.action_delete_entry))
+            },
+            modifier = Modifier.clickable { showDeleteDialog = true },
+            leadingContent = {
+                Icon(
+                    imageVector = Icons.Default.Delete,
+                    contentDescription = null
+                )
+            },
+            colors = ListItemDefaults.colors(
+                headlineColor = MaterialTheme.colorScheme.error,
+                leadingIconColor = MaterialTheme.colorScheme.error,
+                containerColor = Color.Transparent
+            )
+        )
+    }
+}
+
+@Composable
+private fun DeleteDialog(onDismissRequest: () -> Unit, onDeleteEntry: () -> Unit) {
+    AlertDialog(
+        onDismissRequest = onDismissRequest,
+        confirmButton = {
+            TextButton(
+                onClick = onDeleteEntry,
+                colors = ButtonDefaults.textButtonColors(
+                    contentColor = MaterialTheme.colorScheme.error
+                )
+            ) {
+                Text(stringResource(Res.string.action_delete))
+            }
+        },
+        dismissButton = {
+            TextButton(
+                onClick = onDismissRequest
+            ) {
+                Text(stringResource(Res.string.action_cancel))
+            }
+        },
+        title = {
+            Text(stringResource(Res.string.action_delete_entry))
+        },
+        text = {
+            Text(stringResource(Res.string.description_delete_product_entry))
+        }
+    )
 }
